Moved the Home component to the Pages folder. We created a Middleweir folder and moved the AuthenticatedRoute.jsx and UnauthenticatedRoute.jsx files there.
