+++ module-3/web-app/src/Middleware/AuthenticatedRoute.jsx
@@ -1,6 +1,6 @@
 import React, {useContext} from "react"
 import {Navigate, useLocation} from "react-router-dom"
-import {AccountContext} from "./AccountContext"
+import {AccountContext} from "../components/AccountContext"
 
 const AuthenticatedRoute = ({ children }) => {
   const { pathname, search } = useLocation()
