+++ module-3/web-app/src/components/Pages/ForgotPassword/ForgotPassword.module.scss
@@ -1,4 +1,4 @@
-@use '../../mixins/' as *;
+@use '../../../mixins/' as *;
 
 .forgotPasswordPageWrapper {
     display: flex;
