+++ module-3/web-app/src/components/Pages/Home/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import Dashboards from "./Dashboards"
+import Dashboards from "../../Home/Dashboards"
 
 const Home = () => {
     return (
